backend/requirements.txt
asgiref==3.3.0
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.3
django-ckeditor==6.0.0
django-cors-headers==3.5.0
django-heroku==0.3.1
django-js-asset==1.2.2
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.6
pylint==2.6.0
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
whitenoise==5.2.0
wincertstore==0.2
wrapt==1.12.1
